fix(taskworker) Move child process logic into a separate module
By splitting the logic that runs in the spawned child workers, we can
delay importing modules that pull in django. This allows us to defer
loading settings until `configure()` can be called, which allows
getsentry configuration tricks to be applied.
